feat: 팔로우 기능 테스트 및 팔로잉들만 글 조회 기능 (#41)
* feat: follow 기능 구현

* feat: unfollow 기능 구현

* feat: follwers 목록 modal 구현

* fix: follwers 목록 style 수정

* test: testController 생성

* chore: 팀회의 의견 반영

* feat: following list 기능 구현

* test: FeedApiController Test 구현

* test: FeedController Test 구현

* feat: 팔로잉된 사람의 글만 조회 기능 구현

* feat: 팔로잉된 사람의 글만 조회 기능 구현

* fix: 마이페이지 글 목록 순서 날짜 역순으로 수정

+++ src/main/java/com/woowacourse/dsgram/service/facade/Facade.java
@@ -11,6 +11,7 @@ import com.woowacourse.dsgram.service.dto.FollowRelation;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class Facade {
@@ -29,7 +30,9 @@ public class Facade {
         User feedOwner = userService.findByNickName(toNickName);
         long followers =  followService.getCountOfFollowers(feedOwner);
         long followings =  followService.getCountOfFollowings(feedOwner);
-        List<Article> articles = articleService.findArticlesByAuthorNickName(toNickName);
+        List<Article> articles = articleService.findArticlesByAuthorNickName(toNickName)
+                .stream().sorted()
+                .collect(Collectors.toList());
         FollowRelation followRelation = followService.isFollowed(guest,feedOwner);
 
         return FeedInfo.builder()
@@ -44,13 +47,11 @@ public class Facade {
     public void follow(String fromNickName, String toNickName) {
         User guest = userService.findByNickName(fromNickName);
         User feedOwner = userService.findByNickName(toNickName);
-        followService.save(guest,feedOwner);
-    }
 
-
-    public void unfollow(String fromNickName, String toNickName) {
-        User guest = userService.findByNickName(fromNickName);
-        User feedOwner = userService.findByNickName(toNickName);
+        if(!followService.existRelation(guest,feedOwner)) {
+            followService.save(guest,feedOwner);
+            return;
+        }
         followService.delete(guest,feedOwner);
     }
 
@@ -64,6 +65,18 @@ public class Facade {
     public List<FollowInfo> getFollowings(String nickName) {
         User user = userService.findByNickName(nickName);
 
-        return followService.findFollowers(user);
+        return followService.findFollowings(user);
+    }
+
+    public List<Article> getArticlesByFollowings(String nickName) {
+        User user = userService.findByNickName(nickName);
+        List<User> followings = followService.findFollowings(user)
+                .stream().map(followInfo -> userService.findByNickName(followInfo.getNickName()))
+                .collect(Collectors.toList());
+
+        return articleService.findAll()
+                .stream().filter(article -> followings.contains(article.getAuthor()))
+                .sorted()
+                .collect(Collectors.toList());
     }
 }
